Remap random group IDs to check space/RAM saving
randomgroupids is a massive file the script was loading.
It has a set of random group ids for each of the 1000 runs.

Remapping those IDs and separating them into columns is useful.
Each run only needs one column. The rgids could be even generated
on the fly, but then we'd have to treat rgids as an output to
ensure reproducability/error checks.

Remapped random group ID (single run) is 2.5 MB

Comparitively, rdata1.RData is 110 MB, optimized with group IDs
converted to integers is still 85 MB. Clearly large space savings.

This smaller subset of random IDs can be used to create subsets
from the overall dataset on the fly.