Capture and use run stats for each species trends
There are a 1000 runs of Part 3 Step 1. In each run, every species
is processed.  To process each species takes a variable amount of
peak RAM, and CPU time (single thread).  This data may be used to
"schedule" the run order of the computations.

Right now, we sort by time. Species that take the longest to process
are run first.

This may be moot for small region masks, but has potential utility,
especially to tackle cases where the user is running on a low RAM
system.

This requires a previous run of run_species_trends to have generated
stats, which are processed next by gen-species-run-stats.py.
The resulting RData file is packaged inside the docker container
and used.