RAM fill scheduling containter with species subset
Config file can be used to define number of threads and
subset of species to process. run_species_trends_container.R
uses the peakRAM info available from a previous run to
ensure it doesn't exceed available memory. Thus, every
run is guaranteed to succeed - it might take long if the
RAM is less.

Scenarios with less RAM can happen in two ways:
- You are actually on a machine with less RAM
- You're running something like WSL inside Windows.  Windows
  by default allocates some memory, but not all/most of it
  to the container.  You need to adjust that to run more jobs
  in parallel.

The RAM scheduling method isn't actually optimal, but that's
not a requirement for now.