Retry for crashed species jobs
The peakRAM is at best a heurestic.  In edge cases, jobs can get
terminated if there is not enough RAM. I suspect this happens when
one species job is terminating and the other is starting.  But
regardless, the job would terminate at this point.

This change puts the species at the end of the list to be retried
after everything else is done.  A big benefit of this is that
most likely user misconfiguration will never cause the entire run
to fail.  Let's say it's run on a low RAM laptop or machine, then
jobs will get killed but they will be retried later.