data, config, output paths for container
"data" is now in data subdirectory

"configuration" scripts are in "config", under a subdirectory
matching the hostname. For local execution, this is "localhost",
e.g. config/localhost/config.R. For cluster execution, the
hostname will be the hostname of the worker machine.

"output"s are generated in the "output" subdirectory. Again,
output/localhost/{mask}/trends_1.csv for single node, and hostname
changed to worker machine respectively.