Use single large data set+single run random ids
The dataset is thus loaded only once into memory. It is repeatedly
subset for every assignment. For the full country, the dataset is
about 2GB. This thus becomes a constant overhead.

Time to load this dataset is quite small (5-10sec). The dataset
is about 245 MB on disk for the full country.

Considering different number of cores, the per core overhead is:
 - 96 cores < 25 MB
 - 64 cores ~ 35 MB
 - 32 cores ~ 69 MB
 - 16 cores ~ 138 MB

Thus, for any practical cloud run this will not make much
difference.  Flags can be added later to reduce RAM requirement
for local uses on low RAM devices (like laptop)